Port all main classes to PSR-4
