+++ src/PHPTracerWeaver/Transform/BufferEditorInterface.php
@@ -1,0 +1,8 @@
+<?php namespace PHPTracerWeaver\Transform;
+
+use PHPTracerWeaver\Scanner\TokenBuffer;
+
+interface BufferEditorInterface
+{
+    public function editBuffer(TokenBuffer $buffer);
+}
