+++ src/PHPTracerWeaver/Transform/TransformerInterface.php
@@ -1,0 +1,8 @@
+<?php namespace PHPTracerWeaver\Transform;
+
+use PHPTracerWeaver\Scanner\ScannerInterface;
+
+interface TransformerInterface extends ScannerInterface
+{
+    public function getOutput();
+}
