+++ tests/PHPTracerWeaver/DocCommentEditorTransformerTest.php
@@ -1,5 +1,14 @@
 <?php
 
+use PHPTracerWeaver\Scanner\FunctionBodyScanner;
+use PHPTracerWeaver\Scanner\FunctionParametersScanner;
+use PHPTracerWeaver\Scanner\ModifiersScanner;
+use PHPTracerWeaver\Scanner\ScannerMultiplexer;
+use PHPTracerWeaver\Scanner\TokenBuffer;
+use PHPTracerWeaver\Scanner\TokenStreamParser;
+use PHPTracerWeaver\Transform\DocCommentEditorTransformer;
+use PHPTracerWeaver\Transform\MockPassthruBufferEditor;
+use PHPTracerWeaver\Transform\PassthruBufferEditor;
 use PHPUnit\Framework\TestCase;
 
 class TestOfDocCommentEditorTransformer extends TestCase
@@ -31,7 +40,7 @@ class TestOfDocCommentEditorTransformer extends TestCase
         $source = '<?php' . "\n" . 'function bar($x) {}';
         $mock_editor = new MockPassthruBufferEditor();
         $transformer = $this->scan($source, $mock_editor);
-        $this->assertInstanceOf('TokenBuffer', $mock_editor->buffer);
+        $this->assertInstanceOf(TokenBuffer::class, $mock_editor->buffer);
         $this->assertSame('function bar($x) ', $mock_editor->buffer->toText());
     }
 
@@ -40,7 +49,7 @@ class TestOfDocCommentEditorTransformer extends TestCase
         $source = '<?php' . "\n" . 'class Foo { abstract function bar($x) {} }';
         $mock_editor = new MockPassthruBufferEditor();
         $transformer = $this->scan($source, $mock_editor);
-        $this->assertInstanceOf('TokenBuffer', $mock_editor->buffer);
+        $this->assertInstanceOf(TokenBuffer::class, $mock_editor->buffer);
         $this->assertSame('abstract function bar($x) ', $mock_editor->buffer->toText());
     }
 
@@ -57,7 +66,7 @@ class TestOfDocCommentEditorTransformer extends TestCase
         $source = '<?php' . "\n" . '/** Lorem Ipsum */' . "\n" . 'function bar($x) {}';
         $mock_editor = new MockPassthruBufferEditor();
         $transformer = $this->scan($source, $mock_editor);
-        $this->assertInstanceOf('TokenBuffer', $mock_editor->buffer);
+        $this->assertInstanceOf(TokenBuffer::class, $mock_editor->buffer);
         $this->assertTrue($mock_editor->buffer->getFirstToken()->isA(T_DOC_COMMENT));
         $this->assertSame('/** Lorem Ipsum */' . "\n" . 'function bar($x) ', $mock_editor->buffer->toText());
     }
