+++ tests/PHPTracerWeaver/FunctionBodyScannerTest.php
@@ -1,5 +1,7 @@
 <?php
 
+use PHPTracerWeaver\Scanner\FunctionBodyScanner;
+use PHPTracerWeaver\Scanner\TokenStreamParser;
 use PHPUnit\Framework\TestCase;
 
 class TestOfFunctionBodyScanner extends TestCase
