+++ tests/PHPTracerWeaver/FunctionParametersScannerTest.php
@@ -1,5 +1,7 @@
 <?php
 
+use PHPTracerWeaver\Scanner\FunctionParametersScanner;
+use PHPTracerWeaver\Scanner\TokenStreamParser;
 use PHPUnit\Framework\TestCase;
 
 class TestOfFunctionParametersScanner extends TestCase
